requirements.txt
# absl-py==0.9.0
# arabic-reshaper==2.1.3
asgiref==3.4.1
# asn1crypto==1.5.1
# astroid==2.4.2
# astunparse==1.6.3
# async-generator==1.10
# attrs==21.4.0
# autopep8==1.5.4
# autoscraper==1.1.12
# backports.zoneinfo==0.2.1
# beautifulsoup4==4.9.3
# bottle==0.12.19
# bottle-websocket==0.2.9
# bs4==0.0.1
# cachetools==4.1.0
# certifi==2020.6.20
# cffi==1.14.6
# chardet==3.0.4
# charset-normalizer==2.0.12
# click==7.1.2
# colorama==0.4.4
# cryptography==36.0.2
# cssselect2==0.6.0
# cycler==0.11.0
# defusedxml==0.7.1
# diff-match-patch==20200713
# dj-database-url==0.5.0
Django==4.0.4
django-filter==2.4.0
django-import-export==2.5.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.0
# Eel==0.14.0
# et-xmlfile==1.0.1
# excel==1.0.0
# Flask==1.1.2
# fonttools==4.34.4
# future==0.18.2
# gast==0.3.3
# gevent==21.8.0
# gevent-websocket==0.10.1
# google-auth==1.18.0
# google-auth-oauthlib==0.4.1
# google-pasta==0.2.0
# greenlet==1.1.1
# grpcio==1.29.0
gunicorn==20.1.0
# h11==0.13.0
# h5py==2.10.0
html5lib==1.1
# idna==2.9
# isort==5.6.4
# itsdangerous==1.1.0
Jinja2==2.11.2
# Keras-Preprocessing==1.1.2
# kiwisolver==1.4.4
# lazy-object-proxy==1.4.3
# lxml==4.6.3
# Markdown==3.2.2
# MarkupPy==1.14
# MarkupSafe==1.1.1
# matplotlib==3.5.2
# mccabe==0.6.1
# MouseInfo==0.1.3
mysql-connector-python==8.0.21
mysqlclient==2.1.1
numpy==1.18.5
oauthlib==3.1.0
# odfpy==1.4.1
# openpyxl==3.0.7
# opt-einsum==3.2.1
# oscrypto==1.3.0
# outcome==1.1.0
packaging==21.3
# pandas==1.2.3
# pdfkit==1.0.0
Pillow==9.1.0
# plyer==2.0.0
# protobuf==3.12.2
# psycopg2==2.9.3
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# PyAutoGUI==0.9.53
# pycodestyle==2.6.0
# pycparser==2.20
# PyGetWindow==0.0.9
# pyHanko==0.13.1
# pyhanko-certvalidator==0.19.5
# pylint==2.6.0
# pymongo==4.0.1
# PyMsgBox==1.0.9
# PyMySQL==1.0.2
# pyOpenSSL==22.0.0
# pyparsing==2.4.7
# PyPDF3==1.0.6
# pyperclip==1.8.2
# pypng==0.0.21
# PyQRCode==1.2.1
# PyRect==0.1.4
# PyScreeze==0.1.28
# PySocks==1.7.1
# python-bidi==0.4.2
# python-dateutil==2.8.1
# pytube==12.0.0
# pytweening==1.0.4
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
# pywhatkit==5.3
# PyYAML==5.4.1
# qrcode==7.3.1
# reportlab==3.6.9
requests==2.27.1
requests-oauthlib==1.3.0
# rsa==4.6
# scipy==1.4.1
# selenium==4.1.3
# six==1.15.0
# sniffio==1.2.0
# sortedcontainers==2.4.0
# soupsieve==2.2.1
# speedtest-cli==2.1.3
sqlparse==0.2.4
# svglib==1.2.1
# tablib==3.0.0
# tensorboard==2.2.2
# tensorboard-plugin-wit==1.6.0.post3
# tensorflow==2.2.0
# tensorflow-estimator==2.2.0
# termcolor==1.1.0
# tinycss2==1.1.1
# toml==0.10.2
# tqdm==4.64.0
# trio==0.20.0
# trio-websocket==0.9.2
# tzdata==2021.5
# tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
# waitress==2.0.0
# webencodings==0.5.1
# Werkzeug==1.0.1
# whichcraft==0.6.1
whitenoise==5.2.0
# wikipedia==1.4.0
# windows-curses==2.3.0
# wrapt==1.12.1
# wsproto==1.1.0
# xhtml2pdf==0.2.7
# xlrd==2.0.1
# xlwt==1.3.0
# zope.event==4.5.0
# zope.interface==5.4.0
